The collada loader can now load skeleton properly, finally.
Fixed minor bugs related to the texture loading and another bug that didn't allowed models without skeleton to be loaded.
Cleaned up the model folders, fixed the imports for our projects, modified the boost test library to use the proper one for visual studio instead.

+++ nms/nms_utilities/NMS_DebugDraw.h
@@ -16,7 +16,7 @@ static class DEBUGDRAW_D NMS_DebugDraw
 	public:
 		static void drawLine(const Vector& from,const Vector& to,const Vector& fromColor, const Vector& toColor);
 		static void drawLine(const Vector& from,const Vector& to,const Vector& color);
-		static void drawSphere(const Vector& p, int radius, const Vector& color,int lats,int longs);
+		static void drawSphere(const Vector& p, float radius, const Vector& color,int lats,int longs);
 		static void drawBox(const Vector& boxMin, const Vector& boxMax, const Vector& color, int alpha);
 		static void drawTriangle(const Vector& a,const Vector& b,const Vector& c,const Vector& color,int alpha);
 		static void draw3dText(const Vector& location,const char* textString);
